+++ ESA/org.dieschnittstelle.jee.esa.jaxrs/src/org/dieschnittstelle/jee/esa/jaxrs/ITouchpointCRUDWebService.java
@@ -27,7 +27,7 @@ public interface ITouchpointCRUDWebService {
 	
 	@DELETE
 	@Path("/{touchpointId}")
-	public boolean deleteTouchpoint(@PathParam("touchpointId") int id); 
+	public boolean deleteTouchpoint(@PathParam("touchpointId") long id); 
 		
 	/*
 	 * UE JRS1: add a new annotated method for using the updateTouchpoint functionality of TouchpointCRUDExecutor and implement it
