+++ ESA/org.dieschnittstelle.jee.esa.shared.ejbmodule.crm/ejbModule/org/dieschnittstelle/jee/esa/crm/ejbs/crud/CustomerCRUDLocal.java
@@ -9,7 +9,7 @@ public interface CustomerCRUDLocal {
 	
 	public Customer createCustomer(Customer customer);
 
-	public Customer readCustomer(int id);
+	public Customer readCustomer(long id);
 
 	public Customer updateCustomer(Customer customer);
