+++ ESA/org.dieschnittstelle.jee.esa.shared.ejbmodule.crm/ejbModule/org/dieschnittstelle/jee/esa/crm/ejbs/crud/TouchpointCRUDLocal.java
@@ -11,7 +11,7 @@ public interface TouchpointCRUDLocal {
 	
 	public AbstractTouchpoint createTouchpoint(AbstractTouchpoint Touchpoint);
 
-	public AbstractTouchpoint readTouchpoint(int id);
+	public AbstractTouchpoint readTouchpoint(long id);
 
 	public List<AbstractTouchpoint> readAllTouchpoints();
