+++ ESA/org.dieschnittstelle.jee.esa.shared.ejbmodule.erp/ejbModule/org/dieschnittstelle/jee/esa/erp/ejbs/crud/PointOfSaleCRUDLocal.java
@@ -11,9 +11,9 @@ public interface PointOfSaleCRUDLocal {
 	
 	public PointOfSale createPointOfSale(PointOfSale pos);
 
-	public PointOfSale readPointOfSale(int posId);
+	public PointOfSale readPointOfSale(long posId);
 
-	public boolean deletePointOfSale(int posId);
+	public boolean deletePointOfSale(long posId);
 
 	public List<PointOfSale> readAllPointsOfSale();
